% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cfset.R
\name{extractdata}
\alias{extractdata}
\title{Extract from a dataframe all variables used in a formula}
\usage{
extractdata(formula, data, extra = NULL, na.rm = FALSE)
}
\arguments{
\item{formula}{An formula object.}

\item{data}{A dataframe containing all the variables used in \code{formula}.}

\item{extra}{Either a dataframe with additional variables to add to the
output, or a formula object containing additional variables to extract from
\code{data}.}

\item{na.rm}{If \code{TRUE}, listwise delete the newly created dataframe.
Default is \code{FALSE}.}
}
\value{
A data frame.
}
\description{
Extracts from a dataframe all variables used in a formula, saves these
variables to a new dataframe, adds additional variables if requested, and
listwise deletes if desired.  Useful for creating a fully observed dataset
without deleting on unused variables.
}
\author{
Christopher Adolph <\email{cadolph@u.washington.edu}>
}
\keyword{manip}

