% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gator}
\alias{gator}
\title{Alligator food preference}
\format{A data frame with 59 rows and 3 varaibles:
\describe{
\item{food}{Food preference: 1 is Fish, 2 is Invertebrates, 3 is Other.}
\item{size}{Length of the alligator (in meters)}
\item{female}{Is the alligator female?}
}}
\source{
Agresti, A. (2007). An Introduction to Categorical Data Analysis. Wiley
}
\usage{
gator
}
\description{
Data on alligator food preferences.
}
\keyword{datasets}

