% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cfset.R
\name{glue}
\alias{glue}
\title{Paste without separators}
\usage{
glue(...)
}
\arguments{
\item{...}{character strings to be pasted together, or numeric data to be
coerced to character}
}
\description{
Paste without separators
}
\details{
Calls paste with sep="".
}
\author{
Christopher Adolph <\email{cadolph@uw.edu}>
}
\keyword{manip}

