% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{hetnormsimev}
\alias{hetnormsimev}
\title{Simulate expected values from heteroskedastic normal}
\usage{
hetnormsimev(x, b, z, g, ci = 0.95, constant = 1, varconstant = 1,
  predict = TRUE, sims = XXX)
}
\description{
Simulate expected values from heteroskedastic normal
}

