% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{ldvsimfd}
\alias{ldvsimfd}
\title{Simulate lagged dependent variable models out to t periods, first difference}
\usage{
ldvsimfd(x, b, ci = 0.95, constant = 1, xpre = NULL, phi = NULL,
  lagY = NULL, transform = "none", cumulate = FALSE, discount = 0)
}
\description{
Simulate lagged dependent variable models out to t periods, first difference
}

