% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{ldvsimpv}
\alias{ldvsimpv}
\title{Simulate lagged dependent variable models out to t periods, predicted values}
\usage{
ldvsimpv(x, b, ci = 0.95, constant = 1, phi = NULL, lagY = NULL,
  rho = NULL, lagEps = NULL, sigma = NULL, transform = "none",
  initialY = NULL, cumulate = FALSE, discount = 0, nscen = 1)
}
\description{
Simulate lagged dependent variable models out to t periods, predicted values
}

