% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{linearsimfd}
\alias{linearsimfd}
\title{Simulate first difference of expected probabilities for linear models}
\usage{
linearsimfd(x, b, ci = 0.95, constant = 1, xpre = NULL)
}
\description{
Simulate first difference of expected probabilities for linear models
}

