% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{linearsimrr}
\alias{linearsimrr}
\title{Simulate relative risk of expected probabilities for linear models}
\usage{
linearsimrr(x, b, ci = 0.95, constant = 1, xpre = NULL)
}
\description{
Simulate relative risk of expected probabilities for linear models
}

