% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oprobit.r, R/simfns.R
\name{llk.oprobit}
\alias{llk.oprobit}
\title{Likelihood for ordered probit}
\usage{
llk.oprobit(param, x, y)

llk.oprobit(param, x, y)
}
\description{
Likelihood for ordered probit

Likelihood for ordered probit
}

