% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{logitsimrr}
\alias{logitsimrr}
\title{Simulate relative risk of expected probabilities for logit}
\usage{
logitsimrr(x, b, ci = 0.95, constant = 1, xpre = NULL)
}
\description{
Simulate relative risk of expected probabilities for logit
}

