% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{loglinsimfd}
\alias{loglinsimfd}
\title{Simulate first difference of expected counts for a log-linear model}
\usage{
loglinsimfd(x, b, ci = 0.95, constant = 1, xpre = NULL, period = 1,
  labels = NULL)
}
\description{
Simulate first difference of expected counts for a log-linear model
}

