% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeFEdummies.R
\name{makeFEdummies}
\alias{makeFEdummies}
\title{Create a matrix of dummy variables}
\usage{
makeFEdummies(unit, names = NULL)
}
\arguments{
\item{unit}{numeric or character vector of group numbers or names, one for
each observation}

\item{names}{character vector of group names, one for each unique group
number in \code{unit}}
}
\value{
Returns a matrix of binary variables, one row for each observation,
and one column for each group
}
\description{
Creates a matrix of dummy variables for use in model matrix of fixed effects
specifications
}
\author{
Christopher Adolph <\email{cadolph@u.washington.edu}>
}
\keyword{design}

