% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{mlogitsimfd}
\alias{mlogitsimfd}
\title{Simulate first differences for multinomial logit}
\usage{
mlogitsimfd(x, b, ci = 0.95, constant = 1, xpre = NULL, z = NULL,
  zpre = NULL, g = NULL)
}
\description{
Simulate first differences for multinomial logit
}

