% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{mlogitsimrr}
\alias{mlogitsimrr}
\title{Simulate relative risks for multinomial logit}
\usage{
mlogitsimrr(x, b, ci = 0.95, constant = 1, xpre = NULL, z = NULL,
  zpre = NULL, g = NULL)
}
\description{
Simulate relative risks for multinomial logit
}

