% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{oprobitsimfd}
\alias{oprobitsimfd}
\title{Simulate first differences for ordered probit}
\usage{
oprobitsimfd(x, b, ci = 0.95, constant = 1, cat = 3, xpre = NULL)
}
\description{
Simulate first differences for ordered probit
}

