% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{oprobitsimrr}
\alias{oprobitsimrr}
\title{Simulate risk ratio for ordered probit}
\usage{
oprobitsimrr(x, b, ci = 0.95, constant = 1, cat = 3, xpre = NULL)
}
\description{
Simulate risk ratio for ordered probit
}

