% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{probitsimfd}
\alias{probitsimfd}
\title{Simulate first difference of expected probabilities for probit}
\usage{
probitsimfd(x, b, ci = 0.95, constant = 1, xpre = NULL)
}
\description{
Simulate first difference of expected probabilities for probit
}

