% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{probitsimrr}
\alias{probitsimrr}
\title{Simulate relative risk of expected probabilities for probit}
\usage{
probitsimrr(x, b, ci = 0.95, constant = 1, xpre = NULL)
}
\description{
Simulate relative risk of expected probabilities for probit
}

