\name{glue}
\alias{glue}
\title{Paste without separators}
\description{
Paste without separators
}
\usage{
glue(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{character strings to be pasted together, or numeric data to be coerced to character}
}
\details{
 Calls paste with sep="". 
}
\value{
  
}
%\references{ ~put references to the literature/web site here ~ }
\author{Christopher Adolph <\email{cadolph@uw.edu}>}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}



