% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfns.R
\name{llk.oprobit}
\alias{llk.oprobit}
\title{Likelihood for ordered probit}
\usage{
llk.oprobit(param, x, y)
}
\arguments{
\item{param}{Parameters of the ordered probit.}

\item{x}{Matrix of the covariates.}

\item{y}{Vector of the response values.}
}
\description{
Likelihood for ordered probit
}

